Issue#158 user withdrawal (#174)

+++ src/controllers/user.js
@@ -10,7 +10,7 @@ const PersonalSchedule = require('../models/personalSchedule');
 // Error
 const {
   ApiError, DataFormatError,
-  UserNotFoundError, DuplicateUserError, UserIsLeaderError,
+  UserNotFoundError, DuplicateUserError, UserIsLeaderError, BelongToGroupError,
 } = require('../errors');
 
 // Validator
@@ -109,13 +109,17 @@ async function withdrawal(req, res, next) {
       User.findByPk(userId),
       Group.findOne({ where: { leader: userId } }),
     ]);
-
     if (!user) {
       return next(new UserNotFoundError());
     }
 
     if (leader) {
       return next(new UserIsLeaderError());
+    }
+
+    const groupList = await user.getGroups();
+    if (groupList != 0) {
+      return next(new BelongToGroupError());
     }
 
     await PersonalSchedule.destroy({ where: { userId } });
